backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyJWT==2.8.0
python-decouple==3.8
django-filter==23.3
drf-yasg==1.21.7
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.10
requests==2.32.4